🟢feat: Complete YouTube Music Selection Play creation
