+++ app/_/api/type.ts
@@ -13,10 +13,7 @@ export type YouTubeSearchResponse = {
 export type YouTubeSearchItem = {
     kind: string
     etag: string
-    id: {
-        kind: string
-        videoId: string
-    }
+    id: string
     snippet: {
         publishedAt: string
         channelId: string
