+++ app/_components/albumDetail/type.ts
@@ -1,0 +1,52 @@
+import { AlbumType } from "@/app/search/type"
+
+export type AlbumDetailProps = {
+    album: AlbumType
+    onClose: () => void
+}
+
+export type YoutubePlayerProps = {
+    videoId: string
+    onReady?: (event: any) => void
+    onStateChange: (event: any) => void
+}
+
+//
+export type YouTubePlayer = {
+    loadVideoById: (videoId: string) => void
+    playVideo: () => void
+    pauseVideo: () => void
+    setVolume: (volume: number) => void
+}
+export type YouTubeEvent = {
+    target: YouTubePlayer
+    data: number
+}
+export type YouTubePlayerState = {
+    PLAYING: number
+}
+export type YouTubeIframeAPI = {
+    Player: new (
+        elementId: string | HTMLElement,
+        options: {
+            height: string
+            width: string
+            videoId: string
+            events: {
+                onReady: (e: YouTubeEvent) => void
+                onStateChange: (e: YouTubeEvent) => void
+            }
+        },
+    ) => YouTubePlayer
+    PlayerState: YouTubePlayerState
+}
+
+export type AlbumImageProps = {
+    src: string
+    alt: string
+}
+
+export type PlayButtonProps = {
+    togglePlay: () => void
+    isPlaying: boolean
+}
