+++ app/search/type.ts
@@ -15,39 +15,13 @@ export type Thumbnail = {
     high: YouTubeThumbnail
 }
 
-export type YouTubePlayer = {
-    destroy(): void
-    loadVideoById: (videoId: string) => void
-    playVideo: () => void
-    pauseVideo: () => void
-    setVolume: (volume: number) => void
-}
-
-export type YouTubeEvent = {
-    target: YouTubePlayer
-    data: number
-}
-
-export type YouTubePlayerState = {
-    PLAYING: number
-}
-
-export type YouTubeIframeAPI = {
-    Player: new (
-        elementId: string | HTMLElement,
-        options: {
-            height: string
-            width: string
-            videoId: string
-            events: {
-                onReady: (e: YouTubeEvent) => void
-                onStateChange: (e: YouTubeEvent) => void
-            }
-        },
-    ) => YouTubePlayer
-    PlayerState: YouTubePlayerState
-}
-
 export type SearchPageProps = {
     searchParams: { q?: string }
+}
+
+declare global {
+    interface Window {
+        YT: any
+        onYouTubeIframeAPIReady: () => void
+    }
 }
