+++ app/type.ts
@@ -3,13 +3,3 @@ import { ReactNode } from "react"
 export type AppProviderProps = {
     children: ReactNode
 }
-
-type WindowWithYT = Window & {
-    YT: any
-    onYouTubeIframeReady: () => void
-    onYoutubeIframeAPIReady: () => void
-}
-
-declare global {
-    interface Window extends WindowWithYT {}
-}
